Implement a polling mode
This commit adds a third implementation of pwait which uses polling: it
repeatedly checks whether a process with the given ID exists, with
a configurable delay, and exits once it doesn't find such a process.
Polling is less capable than the other two existing modes (netlink and
ptrace) in that it can't capture the exact exit time or the exit code of
the target process, but the benefit is that it doesn't need any special
permissions.

Because polling doesn't capture the target's exit code, I had to use
a different convention for the return code of this implementation. It
will return 0 if a process is found on the first poll, otherwise it
returns -1.