Add CI workflow
This commit adds a continuous integration workflow using Github Actions.
It's pretty simple at this stage, just building the package on Ubuntu
using GCC and Clang. The goal for now is simply to have some kind of
automatic check for compilation errors. In the future I'll add more
platforms, testing, and perhaps uploading of release artifacts.