Add some basic tests of pwait
This commit adds a very simple test framework and a few tests that
verify the functionality of pwait. This is pretty bare-bones for now,
and I expect to expand on this in the future with additional tests and
perhaps a formal test framework, but for now I'm just looking to get
something that checks whether the program basically works on each
commit.