Refactor test code to allow a discrepancy in exit times
I have a test that checks whether pwait and its target process exit at
roughly the same time, which up to this point has been interpreted as
meaning that the clock times are equal to 1-second resolution. However,
the polling mode can't achieve that level resolution; it's possible that
pwait --mode=poll returns multiple seconds after the target process
exits, even if it's functioning properly. So in this commit I'm
refactoring the test code to allow a configurable threshold between when
the target process finishes and when pwait finishes. The threshold
defaults to zero (whole) seconds so that the default behavior will match
what it was before.