Make running the exit code tests optional
To support new waiting methods that don't provide the exit code of
the process being waited for, I'm changing the test script to recognize
an environment variable PWAIT_SKIP_EXIT_CODE_TESTS, which if set to any
nonempty value will cause the exit code tests to be skipped.

Now that the main test function in the test script, run_all_tests(),
supports skipping exit code tests, I changed the test configuration so
that testing the poll method will use that function as well. The tests
of the --delay option, which are specific to the polling method, will be
run from a separate test function.